requirements.txt
# Core dependencies for DinoCompanion Demo
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
pillow>=10.0.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
pydantic>=2.0.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
soundfile>=0.12.0
librosa>=0.10.0
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0 